styling before bed

+++ wwwsrc/src/components/Keep.vue
@@ -6,7 +6,7 @@
     <h5 class="card-title">{{keepData.name}}</h5>
     <p class="card-text">{{keepData.description}}</p>
     <p>{{keepData.views}}, {{keepData.keeps}}</p>
-    <button v-show="this.$auth.user.sub == keepData.userId" class="btn btn-danger" @click="deleteKeep(keepData.id)">x</button>
+    <button v-show="this.$auth.user.sub == keepData.userId" class="btn-small btn-danger" @click="deleteKeep(keepData.id)">x</button>
   </div>
 </div>
 
@@ -30,7 +30,9 @@ export default {
   },
   methods:{
     deleteKeep(keepId){
+      if(window.confirm("Are you sure you want to delete this vault?")){
       this.$store.dispatch("deleteKeep", keepId)
+      }
     },
 
   },
